move NewVideoTile to new file
