+++ src/views/studio/MyVideosView/NewVideoTile/index.ts
@@ -1,0 +1,1 @@
+export * from './NewVideoTile'
